+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Changelog
 
+Version 0.1.1.0
+---------------
+
+*August 21, 2018*
+
+<https://github.com/mstksg/backprop/releases/tag/v0.1.1.0>
+
+*   Added `pointed`, convenient church-style deconstructor for `Pointed` data
+    type.
+*   Added `binFinExt` and `binFinComp`, for "extended" and "compressed"
+    `Finite` indices out of `Bin`s.
+
 Version 0.1.0.0
 ---------------
 
 *August 20, 2018*
+
+<https://github.com/mstksg/backprop/releases/tag/v0.1.0.0>
 
 *   Initial release